requirements.txt
# Formatting / linting modules
nox

# Operational requirements
opencv-contrib-python==4.5.3.56
pyserial==3.5
PyYAML==6.0
simple-repr==1.3.0
RPi.GPIO==0.7.1